fixed logic on comment like as well, along with rightside bar UI improvements
