+++ src/components/postSubComponents/PostComments.js
@@ -1,6 +1,5 @@
 import { v4 as uuidv4 } from "uuid";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faThumbsUp } from "@fortawesome/free-solid-svg-icons";
+import Comment from "./Comment";
 function PostComments(props) {
   return (
     <div className="PostComments">
@@ -18,38 +17,6 @@ function PostComments(props) {
           />
         );
       })}
-    </div>
-  );
-}
-
-function Comment(props) {
-  const handleAddLike = () => {
-    props.handleAddCommentLike(props.postId, props.commentId);
-  };
-
-  return (
-    <div className="Comment">
-      <div className="comment-main">
-        <span className="comment-icon">
-          <img src={props.icon} alt="" />
-        </span>
-        <div className="comment-description">
-          <p className="comment-description-user">{props.userName}</p>
-          <p className="comment-description-text">{props.comment}</p>
-        </div>
-      </div>
-
-      <div className="comment-interaction">
-        <p className="comment-interaction-like" onClick={() => handleAddLike()}>
-          Like
-        </p>
-        <span className="comment-interaction-result">
-          {props.likes > 0 && (
-            <FontAwesomeIcon icon={faThumbsUp} className="result-icon" />
-          )}
-          <p className="result-number">{props.likes > 0 && props.likes}</p>
-        </span>
-      </div>
     </div>
   );
 }
